+++ Harusijack_v.3.1.2/HaruSijak/VM/NewsVM_Fast.swift
@@ -2,7 +2,7 @@ import Foundation
 
 class NewsVM_Fast {
     func loadData(completion: @escaping (Result<[NewsModel_Fast], Error>) -> Void) {
-        guard let url = URL(string: "http://your_server_ip:8000/news/news") else {
+        guard let url = URL(string: "http://127.0.0.1:8000/news/news") else {
             completion(.failure(NSError(domain: "Invalid URL", code: -1, userInfo: nil)))
             return
         }
@@ -12,6 +12,9 @@ class NewsVM_Fast {
                 completion(.failure(error))
                 return
             }
+            if let data = data, let responseString = String(data: data, encoding: .utf8) {
+                   print("Server response: \(responseString)")
+               }
             
             guard let data = data else {
                 completion(.failure(NSError(domain: "No data", code: -1, userInfo: nil)))
@@ -21,7 +24,10 @@ class NewsVM_Fast {
             do {
                 let news = try JSONDecoder().decode([NewsModel_Fast].self, from: data)
                 completion(.success(news))
+                print("**********************************")
+                
                 print(completion(.success(news)))
+                print("**********************************")
             } catch {
                 print("Decoding error: \(error)")
                 completion(.failure(error))
