+++ Harusijack_v.3.1.2/HaruSijak/View/News/NewsView.swift
@@ -25,7 +25,7 @@ import SwiftUI
 struct NewsView: View {
   
     // MARK: * Property *
-    @State var newsList: [NewsModel] = []
+    @State var newsList: [NewsModel_Fast] = []
     @State var searchValue: String = ""
     @FocusState var isTextFieldFocused: Bool
     @State var selectedSegment: String = ""
@@ -33,111 +33,103 @@ struct NewsView: View {
     @State var positive: [String] = ["긍정적 기사", "부정적 기사"]
     let emojis = ["😊", "😭", "🥰", "😠", "😱", "😲"]
     let labels = ["기뻐요", "슬퍼요", "사랑해요", "화나요", "무서워요", "놀라워요"]
-    let percentages = ["80%", "30%", "20%", "15%", "50%", "20%"]
-  
-  var body: some View {
-      
-      
-          ScrollView(content: {
-              VStack(alignment:.leading,content: {
-                  
-                    //검색필드
-//                  HStack(content: {
-//                      Spacer()
-//                      TextField("검색어를 입력하세요.", text: $searchValue)
-//                          .frame(width: UIScreen.main.bounds.width*0.8)
-//                          .textFieldStyle(.roundedBorder)
-//                          .focused($isTextFieldFocused)
-//                      Button(action: {
-//                      }, label: {
-//                          Image(systemName: "magnifyingglass")
-//                              .font(.system(size: 25))
-//                              .foregroundStyle(.black)
-//                      })
-//                      Spacer()
-//                  })
-//                  .padding(.top, 30)
-                  Text("     감정선택")
-                      .bold()
-                      .font(.system(size: 10))
-//                      .font(.custom("Ownglyph_noocar-Rg", size: 40))
-                  Picker("감정 선택 ", selection: $selectedSegment) {
-                      ForEach(0..<emotion.count, id: \.self) { index in
-                          Text(emotion[index])
-                              .tag(index)
-                      }
-                  }
-                  .pickerStyle(SegmentedPickerStyle())
-                  .padding()
-                  Text("     긍정 부정선택(임시)")
-                      .bold()
-                      .font(.system(size: 10))
-                  Picker("감정 선택 ", selection: $selectedSegment) {
-                      ForEach(0..<positive.count, id: \.self) { index in
-                          Text(positive[index])
-                              .tag(index)
-                      }
-                  }
-                  .pickerStyle(SegmentedPickerStyle())
-                  .padding()
-                  
-                  ForEach(newsList, id: \.self) { news in
-                      NavigationLink(destination: NewsArticleView(newslink: news.Link), label: {
-                          VStack(alignment:.leading,content:{
-                              Text(news.Press)
+    
+    var body: some View {
+        ScrollView {
+            VStack(alignment: .leading) {
+                Text("     감정선택")
+                    .bold()
+                    .font(.system(size: 10))
+                
+                Picker("감정 선택 ", selection: $selectedSegment) {
+                    ForEach(0..<emotion.count, id: \.self) { index in
+                        Text(emotion[index])
+                            .tag(index)
+                    }
+                }
+                .pickerStyle(SegmentedPickerStyle())
+                .padding()
+                
+                Text("     긍정 부정선택(임시)")
+                    .bold()
+                    .font(.system(size: 10))
+                
+                Picker("감정 선택 ", selection: $selectedSegment) {
+                    ForEach(0..<positive.count, id: \.self) { index in
+                        Text(positive[index])
+                            .tag(index)
+                    }
+                }
+                .pickerStyle(SegmentedPickerStyle())
+                .padding()
+                
+                ForEach(newsList, id: \.link) { news in
+                    NavigationLink(destination: NewsArticleView(newslink: news.link)) {
+                        VStack(alignment: .leading) {
+                            Text(news.title)
                                 .font(.headline)
                                 .padding([.top, .leading, .trailing])
-//                                .foregroundColor(.primary)
-                              
-                              Text(news.Title)
+                            
+                            Text(news.newContent.prefix(100) + (news.newContent.count > 100 ? "..." : ""))
                                 .font(.subheadline)
                                 .padding([.leading, .bottom, .trailing])
-                                .frame(alignment:.leading)
-//                                .foregroundStyle(Color.gray)
-                              HStack {
-                                  Text("기사 감성예측")
-                                  ForEach(0..<6) { index in
-                                      VStack {
-                                          Spacer()
-                                          
-                                          VStack {
-                                            
-                                              Text(emojis[index])
-                                                  .font(.system(size: 30))
-                                              Text(labels[index])
-                                                  .font(.system(size: 12))
-                                              Text(percentages[index])
-                                                  .font(.system(size: 10))
-                                              Spacer()
-                                          }
-                                          Spacer()
-                                      }
-                                  }
-                              }
-                              .frame(maxWidth: .infinity)
-                          })//VS
+                                .frame(alignment: .leading)
+                            
+                            HStack {
+                                Text("기사 감성예측")
+                                ForEach(0..<6) { index in
+                                    VStack {
+                                        Spacer()
+                                        VStack {
+                                            Text(emojis[index])
+                                                .font(.system(size: 30))
+                                            Text(labels[index])
+                                                .font(.system(size: 12))
+                                            Text(getPercentage(for: index, news: news))
+                                                .font(.system(size: 10))
+                                            Spacer()
+                                        }
+                                        Spacer()
+                                    }
+                                }
+                            }
+                            .frame(maxWidth: .infinity)
+                        }
                         .background(Color.white)
-//                        .shadow(radius: 4)
                         .padding([.top, .horizontal])
-                      })//NVLInk
-                      .buttonStyle(PlainButtonStyle())
-                  }//FE
-              })//VS
-          })//SV
-      
-      .onAppear {
-          let newsVM = NewsVM()
-          newsVM.loadData(url: URL(string: "http://127.0.0.1:5000/news")!) { result in
-              switch result {
-              case .success(let news):DispatchQueue.main.async {
-                  newsList = news
-              }
-              case .failure(let error):
-          print("Error loading news: \(error.localizedDescription)")
-        } // end of switch result 구문
-      } // end of closure newsVM.loadData {result in}
-    } // end of .onAppear
-  } // end of var body: some View
+                    }
+                    .buttonStyle(PlainButtonStyle())
+                }
+            }
+        }
+        .onAppear {
+            let newsVM = NewsVM_Fast()
+            newsVM.loadData { result in
+                switch result {
+                case .success(let news):
+                    DispatchQueue.main.async {
+                        newsList = news
+                    }
+                case .failure(let error):
+                    print("Error loading news: \(error.localizedDescription)")
+                }
+            }
+        }
+    }
+    
+    func getPercentage(for index: Int, news: NewsModel_Fast) -> String {
+        let percentage: Double
+        switch index {
+        case 0: percentage = news.joy
+        case 1: percentage = news.sadness
+        case 2: percentage = news.love
+        case 3: percentage = news.anger
+        case 4: percentage = news.fear
+        case 5: percentage = news.surprise
+        default: percentage = 0
+        }
+        return String(format: "%.0f%%", percentage * 100)
+    }
 } // end of struct NewsView: View
